register detection/sent email working
