+++ webfiles_coutproject/.settings/org.eclipse.wst.common.component
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?><project-modules id="moduleCoreId" project-version="1.5.0">
+    <wb-module deploy-name="cOUT1">
+        <wb-resource deploy-path="/" source-path="/WebContent" tag="defaultRootSource"/>
+        <wb-resource deploy-path="/WEB-INF/classes" source-path="/src"/>
+        <property name="context-root" value="cOUT1"/>
+        <property name="java-output-path" value="/cOUT1/build/classes"/>
+    </wb-module>
+</project-modules>
