+++ webfiles_coutproject/WebContent/TeachersSU.html
@@ -16,7 +16,7 @@
     <h2>Teacher Sign Up</h2>
     <h4>Please fill in this form to create an account.</h4>
     <label for="name"><b>Name</b></label>&nbsp&nbsp<b>(firstname Lastname)</b>
-    <input type = "text" placeholder ="Enter Full Name" id= "Name" required>
+    <input type = "text" placeholder ="Enter Full Name" name= "Name" required>
 
     <label for="email"><b>Email</b></label>&nbsp&nbsp<b>(Please enter Wayne State email address)</b>
     <input type="text" placeholder="Example@wayne.edu" name="email" required pattern = "(\W|^)[\w.+\-]*@wayne\.edu(\W|$)">
