+++ webfiles_coutproject/WebContent/TeachersSU.jsp
@@ -14,7 +14,7 @@ session.putValue("Name",name);
 String pwd=request.getParameter("psw");  
 String email=request.getParameter("email"); 
 Class.forName("com.mysql.jdbc.Driver"); 
-java.sql.Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/cout","root","23paddock");
+java.sql.Connection con = DriverManager.getConnection("jdbc:mysql://localhost:3306/cOUT","root","23paddock");
 Statement st= con.createStatement(); 
 ResultSet rs; 
 int i=st.executeUpdate("insert into Teachers(Name,password,email)values('"+name+"','"+pwd+"','"+email+"')");
